refactor(controller): améliorations de sécu

+++ MonEndoVue.Server/Controllers/BilanQuotidienController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MonEndoVue.Server.Data;
 using MonEndoVue.Server.Models;
+using MonEndoVue.Server.Services;
 using MonEndoVue.Server.ViewModels;
 
 namespace MonEndoVue.Server.Controllers
@@ -10,20 +11,16 @@ namespace MonEndoVue.Server.Controllers
     [Route("[controller]")]
     [ApiController]
     [Authorize]
-    public class BilanQuotidienController(AppDbContext context) : ControllerBase
+    public class BilanQuotidienController(AppDbContext context, CarnetSanteService carnetSanteService) : ControllerBase
     {
-        // GET: BilanQuotidien
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<BilanQuotidien>>> GetBilansQuotidiens()
-        {
-            return await context.BilansQuotidiens.ToListAsync();
-        }
-        
         // GET: BilanQuotidien/5
         [HttpGet("{id}")]
         public async Task<ActionResult<BilanQuotidien>> GetBilanQuotidien(int id)
         {
             var bilanQuotidien = await context.BilansQuotidiens.FindAsync(id);
+            
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, bilanQuotidien?.CarnetSanteId ?? 0);
+            if (securityCheck != null) return securityCheck;
 
             if (bilanQuotidien == null)
             {
@@ -37,6 +34,9 @@ namespace MonEndoVue.Server.Controllers
         [HttpGet("{month}/{year}")]
         public async Task<ActionResult<IEnumerable<BilanQuotidien>>> GetBilanQuotidienByMonth(int carnetSanteId, int month, int year)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, carnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             var bilansQuotidiens = await context.BilansQuotidiens
                 .Where(d => d.Date.Month == month && d.Date.Year == year && d.CarnetSanteId == carnetSanteId)
                 .ToArrayAsync();
@@ -48,6 +48,9 @@ namespace MonEndoVue.Server.Controllers
         [HttpGet("by-week/{carnetSanteId}/{week}/{year}")]
         public async Task<ActionResult<IEnumerable<BilanQuotidien>>> GetBilanQuotidienByWeek(int carnetSanteId, int week, int year)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, carnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             var firstDayOfYear = new DateTime(year, 1, 1);
             var startOfWeek = firstDayOfYear.AddDays((week - 1) * 7 - (int)firstDayOfYear.DayOfWeek + (int)DayOfWeek.Monday);
             var endOfWeek = startOfWeek.AddDays(7);
@@ -65,6 +68,9 @@ namespace MonEndoVue.Server.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutBilanQuotidien(int id, BilanQuotidien bilanQuotidien)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, bilanQuotidien.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             if (id != bilanQuotidien.Id)
             {
                 return BadRequest();
@@ -94,6 +100,9 @@ namespace MonEndoVue.Server.Controllers
         [HttpPost]
         public async Task<ActionResult<BilanQuotidien>> PostBilanQuotidien(BilanQuotidien bilanQuotidien)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, bilanQuotidien.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             context.BilansQuotidiens.Add(bilanQuotidien);
             await context.SaveChangesAsync();
 
@@ -114,6 +123,9 @@ namespace MonEndoVue.Server.Controllers
             {
                 return NotFound();
             }
+            
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, bilanQuotidien.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
 
             context.BilansQuotidiens.Remove(bilanQuotidien);
             await context.SaveChangesAsync();
