+++ MonEndoVue.Server/Controllers/DonneesTransitController.cs
@@ -12,15 +12,9 @@ namespace MonEndoVue.Server.Controllers
     [Authorize]
     public class DonneesTransitController(AppDbContext context, CarnetSanteService carnetSanteService) : ControllerBase
     {
-        // GET: DonneesTransit
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<DonneesTransit>>> GetDonneesTransit()
-        {
-            return await context.DonneesTransit.ToListAsync();
-        }
-        
+       
         // GET: DonneesTransit/5
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<ActionResult<DonneesTransit>> GetDonneesTransit(int id)
         {
             var donneesTransit = await context.DonneesTransit.FindAsync(id);
@@ -29,14 +23,20 @@ namespace MonEndoVue.Server.Controllers
             {
                 return NotFound();
             }
+            
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, donneesTransit.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
 
             return donneesTransit;
         }
         
         // GET: DonneesTransit/ByMonth/5/2021
-        [HttpGet("{carnetSanteId}/{month}/{year}")]
+        [HttpGet("{carnetSanteId:int}/{month:int}/{year:int}")]
         public async Task<ActionResult<IEnumerable<DonneesTransit>>> GetDonneesTransitByMonth(int carnetSanteId, int month, int year)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, carnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             return await context.DonneesTransit
                 .Where(d => d.Date.Month == month && d.Date.Year == year && d.CarnetSanteId == carnetSanteId)
                 .ToArrayAsync();
@@ -47,6 +47,9 @@ namespace MonEndoVue.Server.Controllers
         [HttpPost]
         public async Task<ActionResult<DonneesTransit>> PostDonneesTransit(DonneesTransit donneesTransit)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, donneesTransit.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             context.DonneesTransit.Add(donneesTransit);
             await context.SaveChangesAsync();
 
@@ -58,9 +61,12 @@ namespace MonEndoVue.Server.Controllers
         
         // PUT: DonneesTransit/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> PutDonneesTransit(int id, DonneesTransit donneesTransit)
         {
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, donneesTransit.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
             if (id != donneesTransit.Id)
             {
                 return BadRequest();
@@ -91,7 +97,7 @@ namespace MonEndoVue.Server.Controllers
         }
         
         // DELETE: DonneesTransit/5
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteDonneesTransit(int id)
         {
             var donneesTransit = await context.DonneesTransit.FindAsync(id);
@@ -99,8 +105,11 @@ namespace MonEndoVue.Server.Controllers
             {
                 return NotFound();
             }
-            var carnetSanteId = donneesTransit.CarnetSanteId;
-            carnetSanteService.InvalidateCache(carnetSanteId);
+
+            var securityCheck = await this.ValidateCarnetAccess(carnetSanteService, donneesTransit.CarnetSanteId);
+            if (securityCheck != null) return securityCheck;
+            
+            carnetSanteService.InvalidateCache(donneesTransit.CarnetSanteId);
 
             context.DonneesTransit.Remove(donneesTransit);
             await context.SaveChangesAsync();
